Commit com as ultimas atualizações e alguns testes realizados

+++ Application/Dto/UpdatePedidoItemDto.cs
@@ -5,7 +5,7 @@ namespace GerenciamentoDePedidosWebApi.Application.Dto
     public class UpdatePedidoItemDto
     {
         [Required(ErrorMessage = "O ProdutoId é obrigatório.")]
-        public decimal ProdutoId { get; set; }
+        public int ProdutoId { get; set; }
         [Required(ErrorMessage = "Deve conter pelo menos um produto")]
         public int Quantidade { get; set; }
     }
