+++ Application/Interfaces/IProdutoRepository.cs
@@ -6,7 +6,7 @@ namespace GerenciamentoDePedidosWebApi.Application.Interfaces
 {
     public interface IProdutoRepository : IRepository<Produto>
     {
-        Task<Produto> GetByIdAsync(decimal produtoId);
+        Task<Produto> GetByIdAsync(int produtoId);
         Task<List<Produto>> GetAllAsync();
         Task<dynamic> CadastrarProdutoAsync(Produto produto);
         Task<dynamic> AtualizarProdutoAsync(Produto produto);
